Move Header into todos
and fixup the app bar
