+++ src/todos/index.js
@@ -6,8 +6,6 @@ import sagas from './sagas';
 // re-select
 // import * as selectors from './selectors'
 import MainSection from './components/MainSection';
-import TodoItem from './components/TodoItem';
-import TodoTextInput from './components/TodoTextInput';
 
 export default {
   NAME,
@@ -16,6 +14,4 @@ export default {
   reducer,
   sagas,
   MainSection,
-  TodoItem,
-  TodoTextInput,
 };
